requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
alembic>=1.13
pydantic-settings>=2.2
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
email-validator>=2.1
psycopg[binary]>=3.1
pgvector>=0.2
redis>=5.0
celery>=5.3
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
networkx>=3.1
gunicorn>=20.1

